+++ src/ipor_fusion/PlasmaVaultDataReader.py
@@ -1,5 +1,6 @@
 from dataclasses import dataclass
 
+from eth_abi.exceptions import InsufficientDataBytes
 from web3 import Web3
 
 from ipor_fusion.PlasmaVault import PlasmaVault
@@ -13,10 +14,15 @@ class PlasmaVaultData:
     withdraw_manager_address: str
     asset_address: str
     rewards_claim_manager_address: str
+    price_oracle_middleware_address: str
     fuses: [str]
 
 
 class PlasmaVaultDataReader:
+    ZERO_ADDRESS = Web3.to_checksum_address(
+        "0x0000000000000000000000000000000000000000"
+    )
+
     def __init__(self, transaction_executor: TransactionExecutor):
         self._transaction_executor = transaction_executor
 
@@ -37,6 +43,14 @@ class PlasmaVaultDataReader:
         rewards_claim_manager_address = Web3.to_checksum_address(
             plasma_vault.get_rewards_claim_manager_address()
         )
+        price_oracle_middleware_address = self.ZERO_ADDRESS
+        try:
+            price_oracle_middleware_address = Web3.to_checksum_address(
+                plasma_vault.get_price_oracle_middleware()
+            )
+        except InsufficientDataBytes:
+            pass
+
         fuses = plasma_vault.get_fuses()
         checksum_fuses = [Web3.to_checksum_address(fuse) for fuse in fuses]
         return PlasmaVaultData(
@@ -45,5 +59,6 @@ class PlasmaVaultDataReader:
             withdraw_manager_address=withdraw_manager_address_checksum,
             asset_address=asset_address,
             rewards_claim_manager_address=rewards_claim_manager_address,
+            price_oracle_middleware_address=price_oracle_middleware_address,
             fuses=checksum_fuses,
         )